Draft RPC communication example.
We can now send RPC message via the `window.rpc` (aliased to
`window.external.rpc`) using two public methods:

* `notify()` Call a remote RPC method and immediately resolve.
* `call()` Call a remote method and `await` the returned promise for a reply.

Needs some more work to improve the RPC error handling and
implementations for MacOS and Windows but the basics are in place.

To test it run:

cargo run --example=rpc

And click the `Send Parameters` button, you should see in the terminal
`CONSOLE: Hello, world!` which is the reply from the RPC handler defined
in rpc.rs.