Implement draft RPC API. (#95)
* Implement draft RPC API.

Remove old Callback mechanism.

* Remove obsolete Callback API
* Remove FuncCall and RPC
* Update README
* Rename set_rpc_handler() to set_handler()
* Use shared rpc_proxy() function for platform consistency
* Improve handling of promise cleanup
* Update README with RPC API info.
* Panic if webview handler is set after window creation.
* Improve rpc_proxy() logic, try to ensure any corresponding promise is always removed.
* Remove FuncCall wrapper.

* Remove set_handler().

Use second argument to add_window_with_configs() to set an RpcHandler.

* Fix windows type signature.

* Tidy obsolete comments and code.

* Remove promise fallback clean up code.

So that rust rpc handlers can perform asynchronous tasks and defer
promise evaluation until later.

If an rpc handler returns None then the handler takes responsibility for
ensuring the corresponding promise is resolved or rejected. If the
request contains an `id` then the handler *must* ensure it evaluates
either `RpcResponse::into_result_script()` or
`RpcResponse::into_error_script()`.

* Remove Sync bound from RpcHandler.

Update multiwindow example so it is slightly more illustrative of
a real-world use case. Now it launches a window when a button is clicked
in the main webview. Multiple windows can be launched and the URL for
the new window is passed from the Javascript code.

* Remove urlencoding from examples.