Add `Application` back, require it for `WebviewBuilder`
Linux only so far. Until further notice, this entire branch is going to only
work on Linux until I can decide on a good interface to expose WebDriver stuff.

`Application` makes it so that multiple windows can share a single WebContext
and UserContentManager (data directory). This is required for WebDriver because
only a single `WebContext` is allowed to be marked to allow automation. It
seems that macOS and Windows have similar limitations to some extent.

Only the multi_window example has been updated so far.