Rework example
Now holds temp_dir as long as necessary to prevent premature drop (and
thus deleting our example file). Requires the use of the Rc<RefCell>
pattern for interior mutability across a shared reference - this
(should) be safe, as the entire example is single threaded, and any
mutable borrows should be dropped by the time a borrow ocurrs.