feat: add MacOS print options
Currently, when printing on a MacOS, it defaults to a quite big margin, which does not reflect the side of
it when printing on the web. The current code just creates a new NSPrintInfo from the default, which for some
reason has a large margin.

This changes so that the default margin is 0 and add the option to change it when printing the webview on MacOs.
This is not a field that can be changed by the user on the appkit modal, so it has to be defined before the modal is open.

Other options could be added on top of these ones, but for this initial commit/PR decided to go with adding the margins
since it seems to be the one with bigger discrepancy compared to the web version.

References:
- https://developer.apple.com/documentation/appkit/nsprintinfo