fix(linux): don't share UserContentManager between webviews
Fixes #1308

Init scripts are registered to the UserContentManager, so if it is
shared between webviews, the init scripts will conflict. Use a unique
UserContentManager for each WebView instead of sharing it across the
WebContext.

The IPC message handler is also registered to UserContentManager.
window_hash is no longer needed as a key to
register_script_message_handler since the manager is unique. window_hash
is already removed on the dev branch.