fix(linux): don't try to register URI scheme if it already exists
Change from register_uri_scheme to try_register_uri_scheme, which
doesn't attempt to register if we already know that one with the same
name was registered to the WebContext before. Without this change, I get
this error from WebKitGTK:

** (tauri-app:[...]): CRITICAL **: 14:03:50.040: Cannot register URI scheme tauri more than once

It doesn't seem to affect functionality, but better to issue a more
appropriate warning on our side.