Windows: Use `CoreWebView2EnvironmentOptions` API.
Use the CoreWebView2EnvironmentOptions API wrapper when creating a
WebView2 environment on Windows, as opposed to using the
ICoreWebView2EnvironmentOptions* COM interfaces. The wrapper exposes
the interface methods with a more Rust-like API.

It may be possible to do this with ICoreWebView2Controller as well, but
I don't feel like experimenting with that right now. This is just an
improvement I realised I could make while implementing #1339.

Signed-off-by: Jamie Ridding <jamie.ridding@hotmail.co.uk>