feat(windows): add `WebViewBuilderExtWindows::with_scroll_bar_style` (#1344)
* Windows: Use `CoreWebView2EnvironmentOptions` API.

Use the CoreWebView2EnvironmentOptions API wrapper when creating a
WebView2 environment on Windows, as opposed to using the
ICoreWebView2EnvironmentOptions* COM interfaces. The wrapper exposes
the interface methods with a more Rust-like API.

It may be possible to do this with ICoreWebView2Controller as well, but
I don't feel like experimenting with that right now. This is just an
improvement I realised I could make while implementing #1339.

Signed-off-by: Jamie Ridding <jamie.ridding@hotmail.co.uk>

* Add with_scroll_bar_style to WebViewBuilderExtWindows

Signed-off-by: Jamie Ridding <jamie.ridding@hotmail.co.uk>

* Add change file.

Signed-off-by: Jamie Ridding <jamie.ridding@hotmail.co.uk>

* Adjust change type in change file.

I used the wrong change type for wry in 1344-with_scroll_bar_style.
This commit upgrades the change type from a patch to a minor change.

Signed-off-by: Jamie Ridding <jamie.ridding@hotmail.co.uk>

* I forgot to run cargo fmt -_-

Signed-off-by: Jamie Ridding <jamie.ridding@hotmail.co.uk>

---------

Signed-off-by: Jamie Ridding <jamie.ridding@hotmail.co.uk>