fix(linux): don't share UserContentManager between webviews (v0.24) (#1325)
Fixes #1308

Init scripts are registered to the `UserContentManager`, so if it is shared between `WebView`s, the init scripts conflict. Use a unique `UserContentManager` for each `WebView` instead of sharing it across the `WebContext`.

The IPC message handler is also registered to `UserContentManager`.

Also changed from `register_uri_scheme` to `try_register_uri_scheme`, which doesn't attempt to register a URI scheme if we already know that one with the same name was registered to the `WebContext` before. Without this change, I get this error from WebKitGTK:
```
** (tauri-app:[...]): CRITICAL **: 14:03:50.040: Cannot register URI scheme tauri more than once
```
It doesn't seem to affect functionality, but better to issue a more appropriate warning on our side. (On dev, this is propagated as an error)